+++ crates/aptos/src/update/update_helper.rs
@@ -48,7 +48,10 @@ pub fn build_updater(
     let target = format!("{}-{}", arch_str, target);
 
     let install_dir = match install_dir.clone() {
-        Some(dir) => dir,
+        Some(dir) => {
+            println!("dir:{:?}", dir);
+            dir
+        },
         None => {
             let dir = get_additional_binaries_dir();
             // Make the directory if it doesn't already exist.
@@ -63,6 +66,12 @@ pub fn build_updater(
         None => "0.0.0",
     };
 
+    println!(
+        "install_dir:{:?}, is directory:{}",
+        install_dir,
+        install_dir.is_dir()
+    );
+
     Update::configure()
         .bin_install_dir(install_dir)
         .bin_name(binary_name)
@@ -75,7 +84,13 @@ pub fn build_updater(
         .map_err(|e| anyhow!("Failed to build self-update configuration: {:#}", e))
 }
 
-pub fn get_path(name: &str, exe_env: &str, binary_name: &str, exe: &str) -> Result<PathBuf> {
+pub fn get_path(
+    name: &str,
+    exe_env: &str,
+    binary_name: &str,
+    exe: &str,
+    find_in_path: bool,
+) -> Result<PathBuf> {
     // Look at the environment variable first.
     if let Ok(path) = std::env::var(exe_env) {
         return Ok(PathBuf::from(path));
@@ -87,9 +102,11 @@ pub fn get_path(name: &str, exe_env: &str, binary_name: &str, exe: &str) -> Resu
         return Ok(path);
     }
 
-    // See if we can find the binary in the PATH.
-    if let Some(path) = pathsearch::find_executable_in_path(exe) {
-        return Ok(path);
+    if find_in_path {
+        // See if we can find the binary in the PATH.
+        if let Some(path) = pathsearch::find_executable_in_path(exe) {
+            return Ok(path);
+        }
     }
 
     Err(anyhow!(